+++ installer/fileserver/tasks/tagvm/tagvm.go
